backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
spacy==3.7.2
python-dotenv==1.0.0
